fixed ci tests
